backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
numpy
supabase
python-dotenv
